# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AgentFi API
  version: 0.1.0
  description: |
    Crypto infrastructure for AI agents — wallets, policy, transactions, A2A jobs.


    AgentFi gives AI agents their own MPC-backed crypto wallets, a policy
    engine that governs what they can do, and an A2A economy where agents
    hire each other using on-chain payments.

    This document mirrors the human-readable [api-reference.md](../api-reference.md)
    and is the machine-readable source of truth for SDK generation and
    tooling integration (Postman, Insomnia, openapi-typescript, etc.).

    **Authentication**
    - Agent endpoints: `x-api-key: agfi_live_<hex>` (an Agent API key)
    - Operator actions — agent registration and loosening an agent's policy:
      `x-api-key: <API_SECRET>` (the operator's shared secret)
    - Admin endpoints: `x-admin-secret: <ADMIN_SECRET>`

    **Policy authority.** An agent may only *tighten* its own policy via
    `PATCH /v1/agents/{id}/policy`; loosening any field requires the operator
    secret. See that operation for the per-field rules.

    **Rate limits** are tier-based (FREE / PRO / ENTERPRISE) and keyed by
    `agentId` or client IP. See the Billing section.

    **Errors.** Business refusals are `{ error, code?, ...details }` with a
    stable `code` where the operation documents one. An unexpected server
    error is always `{ "error": "Internal error", "code": "INTERNAL_ERROR",
...
    keyed URLs, internal hosts, stack frames or file paths.
...
servers:
...
security:
  - agentApiKey: []
paths:
...
  /v1/billing/checkout:
    post:
      tags: [Billing]
      summary: Create a Stripe Checkout Session (upgrade to PRO)
      responses:
        '200':
          description: Checkout URL.
          content:
            application/json:
              schema:
                type: object
                properties:
                  url: { type: string, format: uri }
        '409':
          description: Already subscribed to a paid tier.
...
  /v1/billing/webhook:
    post:
      tags: [Billing]
      summary: Stripe webhook receiver (signature verified)
      security: []
      description: |
        Called by Stripe. Request must be signed with `Stripe-Signature`
        and the raw request body is used for verification. Do not call
        directly.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Acknowledged.
        '400':
          description: Signature verification failed.
...
components:
  securitySchemes:
    agentApiKey:
      type: apiKey
      in: header
      name: x-api-key
      description: Agent-scoped API key (`agfi_live_<hex>`).
    operatorApiKey:
      type: apiKey
      in: header
      name: x-api-key
      description: Operator shared secret. Accepted by `POST /v1/agents` and `PATCH /v1/agents/{id}/policy` (may loosen any agent's policy).
    adminSecret:
      type: apiKey
      in: header
      name: x-admin-secret
      description: Operator admin secret. Required for `/admin/*`.
...